Feature/usage and monitoring

+++ mq-spec/src/main/java/at/uastw/disys26bwi/mqSpec/dto/EnergyNodeMessageDto.java
@@ -4,11 +4,12 @@ import at.uastw.disys26bwi.mqSpec.constants.Association;
 import at.uastw.disys26bwi.mqSpec.constants.NodeType;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public record EnergyNodeMessageDto(
   NodeType type,
   Association association,
-  double kwh,
+  BigDecimal kwh,
   String datetime
 ) implements Serializable {
 }
